testing/web-platform/tests/shape-detection/detection-HTMLVideoElement-invalid-state.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>

const videoElementTests =
    [
      {
        createDetector: () => { return new FaceDetector(); },
        name: "Face - detect(HTMLVideoElement)",
      },
      {
        createDetector: () => { return new BarcodeDetector(); },
        name: "Barcode - detect(HTMLVideoElement)",
      }
    ];

for (let videoElementTest of videoElementTests) {

  // Detector's detect() rejects on a HAVE_NOTHING HTMLVideoElement.
  promise_test(async t => {
    const video = document.createElement("video");
    video.src = "";
    const videoWatcher = new EventWatcher(t, video, ["play", "error"]);
    video.load();
    await videoWatcher.wait_for("error");
    assert_equals(video.readyState, video.HAVE_NOTHING);

    const detector = videoElementTest.createDetector();
    await promise_rejects(t, 'InvalidStateError', detector.detect(video));
  }, `${videoElementTest.name} - HAVE_NOTHING`);

  // Detector's detect() rejects on a HAVE_METADATA HTMLVideoElement.
  promise_test(async t => {
    const url = '/media/test-av-384k-44100Hz-1ch-320x240-30fps-10kfr.webm';
    const type = 'video/webm; codecs="vp8, vorbis"';
    const video = document.createElement("video");
    document.body.appendChild(video);
    t.add_cleanup(() => { document.body.removeChild(video); });

    // Attach MediaSource to the video element.
    const mediaSource = new MediaSource();
    const mediaSourceURL = URL.createObjectURL(mediaSource);
    const mediaSourceWatcher =
        new EventWatcher(t, mediaSource, ["sourceopen", "error"]);
    video.src = mediaSourceURL;
    await mediaSourceWatcher.wait_for("sourceopen");
    const sourceBuffer = mediaSource.addSourceBuffer(type);

    // Load media data from the video file to create an initialization segment.
    const response = await fetch(url);
    const mediaData = await response.arrayBuffer();
    const initSegment = new Uint8Array(mediaData, 0, 4052);

    // Append the initialization segment to trigger a transition
    // to HAVE_METADATA.
    const videoWatcher =
        new EventWatcher(t, video, ["loadedmetadata", "error"]);
    sourceBuffer.appendBuffer(initSegment);

    await videoWatcher.wait_for("loadedmetadata");
    assert_equals(video.readyState, video.HAVE_METADATA);

    const detector = videoElementTest.createDetector();
    await promise_rejects(t, 'InvalidStateError', detector.detect(video));
  }, `${videoElementTest.name} - HAVE_METADATA`);

}

</script>
